Update t's Formula test
t is designed to be interactive and open tmux. This isn't possilble in a
non-interactive environment like CI.

From the [Homebrew Formula Cookbook][1]:

> We want tests that don’t require any user input and test the basic
functionality of the application. For example `foo build-foo input.foo`
is a good test and (despite their widespread use) `foo --version` and
`foo --help` are bad tests. However, a bad test is better than no test
at all.

So yeah, I know this is a bad test.

[1]: https://docs.brew.sh/Formula-Cookbook#add-a-test-to-the-formula